refactor: ShowReview getList Query 로 분리
