+++ web/src/main/java/org/cmc/curtaincall/web/common/response/ListResult.java
@@ -1,0 +1,8 @@
+package org.cmc.curtaincall.web.common.response;
+
+import java.util.List;
+
+public record ListResult<T>(
+        List<T> content
+) {
+}
