+++ web/src/test/java/org/cmc/curtaincall/web/review/ShowReviewControllerDocsTest.java
@@ -4,15 +4,11 @@ import org.cmc.curtaincall.web.common.AbstractWebTest;
 import org.cmc.curtaincall.web.common.response.IdResult;
 import org.cmc.curtaincall.web.review.request.ShowReviewCreate;
 import org.cmc.curtaincall.web.review.request.ShowReviewEdit;
-import org.cmc.curtaincall.web.review.response.ShowReviewResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.data.domain.SliceImpl;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-
-import java.util.List;
 
 import static org.cmc.curtaincall.web.common.RestDocsAttribute.constraint;
 import static org.mockito.ArgumentMatchers.any;
@@ -22,7 +18,8 @@ import static org.springframework.restdocs.headers.HeaderDocumentation.requestHe
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.*;
 import static org.springframework.restdocs.payload.PayloadDocumentation.*;
-import static org.springframework.restdocs.request.RequestDocumentation.*;
+import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
+import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -66,59 +63,6 @@ class ShowReviewControllerDocsTest extends AbstractWebTest {
                         ),
                         responseFields(
                                 fieldWithPath("id").description("공연 리뷰 ID")
-                        )
-                ));
-    }
-
-    @Test
-    void getList_Docs() throws Exception {
-        // given
-        List<ShowReviewResponse> reviewResponseList = List.of(
-                ShowReviewResponse.builder()
-                        .id(5L)
-                        .showId("PF223355")
-                        .grade(4)
-                        .content("좋아요")
-                        .creatorId(4L)
-                        .creatorNickname("고라파덕")
-                        .creatorImageUrl("http://image-url")
-                        .build()
-        );
-        given(showReviewService.getList(any(), any()))
-                .willReturn(new SliceImpl<>(reviewResponseList));
-
-        // expected
-        mockMvc.perform(get("/shows/{showId}/reviews", "PF220846")
-                        .header(HttpHeaders.AUTHORIZATION, "Bearer ACCESS_TOKEN")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .accept(MediaType.APPLICATION_JSON)
-                        .param("page", "0")
-                        .param("size", "20")
-                )
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andDo(document("show-review-get-list",
-                        requestHeaders(
-                                headerWithName(HttpHeaders.AUTHORIZATION).description("인증 필요")
-                        ),
-                        queryParameters(
-                                parameterWithName("page").description("페이지 번호"),
-                                parameterWithName("size").description("페이지 사이즈").optional()
-                        ),
-                        pathParameters(
-                                parameterWithName("showId").description("공연 ID")
-                        ),
-                        responseFields(
-                                beneathPath("content[]").withSubsectionId("content"),
-                                fieldWithPath("id").description("공연 리뷰 ID"),
-                                fieldWithPath("showId").description("공연 ID"),
-                                fieldWithPath("grade").description("평점"),
-                                fieldWithPath("content").description("리뷰 내용"),
-                                fieldWithPath("creatorId").description("작성자 ID"),
-                                fieldWithPath("creatorNickname").description("작성자 닉네임"),
-                                fieldWithPath("creatorImageUrl").description("작성자 프로필 이미지").optional(),
-                                fieldWithPath("createdAt").description("생성일시"),
-                                fieldWithPath("likeCount").description("좋아요 개수")
                         )
                 ));
     }
